# Conflicts:
#	scripts/startup/bl_ui/node_add_menu_geometry.py
#	source/blender/blenkernel/BKE_node.h
#	source/blender/makesrna/intern/rna_nodetree.cc
#	source/blender/nodes/geometry/node_geometry_register.cc
#	source/blender/nodes/geometry/node_geometry_register.hh

+++ intern/mikktspace/mikk_float3.hh
@@ -1,4 +1,4 @@
-/* SPDX-FileCopyrightText: 2020-2023 Blender Foundation
+/* SPDX-FileCopyrightText: 2020-2023 Blender Authors
  *
  * SPDX-License-Identifier: Apache-2.0 */
 
@@ -64,7 +64,7 @@ struct float3 {
 
   float length() const
   {
-    return sqrt(length_squared());
+    return sqrtf(length_squared());
   }
 
   static float distance(const float3 &a, const float3 &b)
